backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
APScheduler==3.10.4
astroid==3.0.2
attrs==23.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
black==23.12.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
dataclasses-json==0.6.3
dill==0.3.7
distro==1.9.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.1
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Migrate==2.6.0
Flask-Script==2.0.5
Flask-SQLAlchemy==3.1.1
flup==1.0.3
frozenlist==1.4.1
google-api-core==2.15.0
google-api-python-client==2.114.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
idna==3.6
isort==5.13.2
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.352
langchain-community==0.0.11
langchain-core==0.1.9
langsmith==0.0.79
lxml==5.1.0
mailtrap==2.0.1
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.2
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.3
openai==0.28.0
packaging==23.2
pandas==2.1.4
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
protobuf==4.25.2
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pylint==3.0.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-file==1.5.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
setuptools==68.2.2
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
sqlmodel==0.0.14
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.2
tinysegmenter==0.3
tldextract==5.1.1
tomlkit==0.12.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.1.0
Werkzeug==3.0.1
wheel==0.41.2
yarl==1.9.4
